requirements.txt
discord.py
hercai